diagnotix/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
google-generativeai>=0.8.0
requests
pandas
networkx
python-dotenv
pyvis
